use higher quality upper bits for mult-hashing finalization
the upper bits of w-bit multiplications are of higher quality regarding
hashing, i.e., they are more dependent on the input number.
It is therefore preferable to consume them instead of the lower bits
when hashtable size < 2^32. To prevent accidents, this hash finalization
step is now directly implemented in the mult_hash functions.

* add docstrings
* deduplicate TODOs